+++ R/globals.R
@@ -384,6 +384,8 @@ utils::globalVariables(c(
   "match_status",
   "clean_name",
   "pos_group",
+  "pos_grp",
+  "panna_gk_scaled",
   "modal_pos",
   "goals_scored",
   "goals_conceded",